+++ util/opentelemetry-util-genai/CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3944](#3944))
 - Add metrics to LLMInvocation traces
   ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3891](#3891))
+- Add parent class genAI invocation
+  ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3889](#3889))
 
 ## Version 0.2b0 (2025-10-14)